+++ src/Transport/Messages/ClientHandlers/ClientCore.cs
@@ -1,0 +1,27 @@
+﻿/**
+ * Copyright (C) 2020 Quaint Studios, Kristopher Ali (Makosai) <kristopher.ali.dev@gmail.com>
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Affero General Public License as
+ * published by the Free Software Foundation, either version 3 of the
+ * License, or (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU Affero General Public License for more details.
+ *
+ * You should have received a copy of the GNU Affero General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
+namespace Sustenet.Transport.Messages.ClientHandlers
+{
+    /// <summary>
+    /// The core of the client's messaging system. Unused currently.
+    /// </summary>
+    static class ClientCore
+    {
+
+    }
+}
