+++ src/Transport/Messages/ClientHandlers/ClientSend.cs
@@ -15,13 +15,13 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-namespace Sustenet.Transport.Messages
+namespace Sustenet.Transport.Messages.ClientHandlers
 {
     using Network;
     using Clients;
 
     /// <summary>
-    /// 
+    /// TODO: Documentation
     /// </summary>
     static class ClientSend
     {
@@ -46,43 +46,6 @@ namespace Sustenet.Transport.Messages
             {
                 BaseClient.DebugClient(client.id, "Cannot login unless connected to a Master Server.");
             }
-        }
-    }
-
-    /// <summary>
-    /// Messages that the Client would receive from a server.
-    /// </summary>
-    static class ClientReceive
-    {
-        /// <summary>
-        /// Handle a message from the server.
-        /// </summary>
-        /// <param name="client">The client who received the message.</param>
-        /// <param name="packet">The packet containing the message from the server.</param>
-        internal static void Message(this Client client, Packet packet)
-        {
-            string msg = packet.ReadString();
-
-            BaseClient.DebugClient(client.id, $"(Server Message) {msg}");
-        }
-
-        /// <summary>
-        /// Initializes the client's ID and username.
-        /// If the client is a Cluster, the username is the key.
-        /// 
-        /// TODO: Change to the cluster config name in the future.
-        /// </summary>
-        /// <param name="client">The client whose ID and username should be changed.</param>
-        /// <param name="packet">The packet containing the new client ID.</param>
-        internal static void InitializeClient(this Client client, Packet packet)
-        {
-            string username = packet.ReadString();
-            int id = packet.ReadInt();
-
-            client.name = username;
-            client.id = id;
-
-            BaseClient.DebugClient(client.id, $"Welcome, {username}!");
         }
     }
 }
